Refactor feedback and assigned mentee paths
