+++ src/features/Feedback/index.ts
@@ -1,1 +1,3 @@
 export { default as MentorQuarterlyFeedback } from './MentorQuarterlyFeedback'
+export { default as MenteeFeedback } from './MenteeFeedback'
+export { default as MentorFeedback } from './MentorFeedback'
